RobotGridTests.java
import org.junit.Assert;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

public class RobotGridTests {

    RobotGrid robotGrid = new RobotGrid();

    @Test
    public void jankBacktrack_test1() {
        var actual = robotGrid.jankBacktrack(23, 23);
        assertTrue(actual);
    }

    @Test
    public void testIsSafe() {
        var actual = robotGrid.isSafe(123,123);
        assertFalse(actual);
    }

    @Test
    public void testIsSafeFalse() {
        var actual = robotGrid.isMined(12345667,-1236677);
        assertTrue(actual);
    }

    @Test
    public void testIsMined() {
        var actual = robotGrid.isMined(123,123);
        assertFalse(actual);
    }

    @Test
    public void testIsMinedFalse() {
        var actual = robotGrid.isMined(12345667,-1236677);
        assertTrue(actual);
    }

    @Test
    public void testGetSquarePositiveValue() {
        int value = robotGrid.getSquareValue(1234);
        assertEquals(value, 10);
    }

    @Test
    public void testGetSquareNegativeValue() {
        int value = robotGrid.getSquareValue(-12345);
        assertEquals(value, 15);
    }

    @Test
    public void testGetSquareZeroValue() {
        int value = robotGrid.getSquareValue(0);
        assertEquals(value, 0);
    }

    @Test
    public void testTotalSafeSquares() {
        // TODO: Add test logic and assertions for totalSafeSquares function
        Assert.fail("Not implemented yet");
    }

    @Test
    public void testShortestSafeJourney() {
        // TODO: Add test logic and assertions for shortestSafeJourney function
        Assert.fail("Not implemented yet");
    }


    // TODO: Add more tests to ensure comprehensive test coverage
}
